requirements.txt
streamlit
langchain
langgraph
langchain-openai
langchain-community
openai
sentence-transformers
transformers
torch
pillow
pandas
neo4j
weaviate-client
tavily-python
pytesseract
spacy
faiss-cpu
python-dotenv
accelerate
